FIxed crash caused by returning to the app after a while
This crash occurs because over time the settings that were loaded in will be deleted from memory. When the user returns to the app, and the activity left off was anything other than the deck activity, the app will immediately crash. This bug was fixed by calling the loadData() method on any activities that require the data to be loaded.